Ajouter attaque distance pirate
Ajout script d'attaque a distance, setup d'un prefab "Pirate" qui bouge et tire. Ajout d'une scene "Test Attaque" pour tester les dégâts.
Encore rien en réseau pour les dégâts, il n'y a que le tire a distance, mais il y a des dégâts en AOE, et normalement en local ca marche avec les barres de vie
Changement dans "HealthScript" ->changement de int a float pour les valeurs de vie / dégâts.

+++ Assets/Scripts/HealthScript.cs
@@ -8,7 +8,7 @@ namespace PUNTutorial
     public class HealthScript : Photon.PunBehaviour
     {
         public Slider HealthSlider;
-        private int HealthMax;
+        private float HealthMax;
 
         public void Start()
         {
@@ -19,7 +19,7 @@ namespace PUNTutorial
             HealthSlider.value = HealthMax;
         }
 
-        public void Damage(int damage)
+        public void Damage(float damage)
         {
             HealthSlider.value = HealthSlider.value - damage;
         }
